+++ src/main/java/com/alkemy/ong/repositories/INewsRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repositories;
 
-import com.alkemy.ong.models.News;
+import com.alkemy.ong.entities.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
