+++ src/main/java/com/alkemy/ong/repositories/IRoleRepository.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.repositories;
 
-import com.alkemy.ong.models.Role;
+import com.alkemy.ong.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
